Use long instead of int to avoid casting

+++ src/main/java/org/opensearch/index/store/niofs/CryptoBufferedIndexInput.java
@@ -165,9 +165,9 @@ final class CryptoBufferedIndexInput extends BufferedIndexInput {
         }
 
         // Calculate frame boundaries using bit operations (frameSize is power of 2)
-        int frameNumber = (int)(position >>> frameSizePower);
+        long frameNumber = position >>> frameSizePower;
         long offsetWithinFrame = position & ((1L << frameSizePower) - 1);
-        long frameEnd = ((long) frameNumber + 1) << frameSizePower;
+        long frameEnd = (frameNumber + 1) << frameSizePower;
 
         // Limit read to not cross frame boundary
         int maxReadInFrame = (int) Math.min(dst.remaining(), frameEnd - position);
